scripts/requirements.txt
psycopg2-binary>=2.9
python-dotenv>=1.0
requests>=2.31
pandas>=2.1
